Experimental new way of consuming excess newlines
Removes the concept of left/right glue that were previously added to all inline logic, and instead automatically chomps through any whitespace at the start and end of a function call. Main reason is that left/right glue was leaking when you divert out of a piece of inline logic. This is possibly/probably more elegant, but not sure if there are weird edge cases. Tests are passing.